Перенос бизнес-логики в отдельный файл
